Fixed and improved tests for the address component
